2019/3/18   18：47 HY  添加删除逻辑和投诉建议接口

+++ business/biz/proposal.js
@@ -7,7 +7,7 @@ let dao = require("../../db_config/dao"),
     config = require('../../db_config/config'),
     exception = require('../../utils/exception.js'),
     str = require("../../utils/stringHelper"),
-    proxyDao = require('../../business/dao/proxy.js'),
+    proposalDao = require('../../business/dao/proposal.js'),
     deleteDao = require('../../business/dao/delete.js'),
     sysDao = require('../../business/dao/sysConfig.js');
 
@@ -15,11 +15,11 @@ var uuid = require('node-uuid');
 
 class biz {
     //查询
-    static async proxys(params) {
+    static async proposals(params) {
         return await dao.manageConnection(async (connection) => {
             //获取活动主信息
-            var result = await proxyDao.queryProxy(connection, params);
-            var resultCount = await proxyDao.queryProxyCount(connection, params);
+            var result = await proposalDao.queryProposal(connection, params);
+            var resultCount = await proposalDao.queryProposalCount(connection, params);
             let count = 0;
             if (resultCount && resultCount.length > 0) {
                 count = resultCount[0].count;
@@ -33,22 +33,25 @@ class biz {
     //添加申请
     static async add(params) {
         return await dao.manageConnection(async (connection) => {
-            return await proxyDao.addProxy(connection,params)
+            return await proposalDao.addProposal(connection,params)
         })
     }
 
     //修改申请
     static async update(params) {
         return await dao.manageConnection(async (connection) => {
-            return await proxyDao.updateProxy(connection,params)
+            return await proposalDao.updateProposal(connection,params)
         })
     }
 
-    //删除代理
+    //删除代理申请
     static async delete(params) {
         return await dao.manageConnection(async (connection) => {
-            let param = {tableName: "user_proxy_apply", fieldName: "id", fieldValue: params.id}
-            return await deleteDao.delete(connection, param);
+            let result = await proposalDao.queryProposal(connection, params);
+            if(result && result.length > 0){
+                let param = {tableName: "user_proposal", fieldName: "id", fieldValue: params.id}
+                return await deleteDao.delete(connection, param);
+            }
         })
     }
 }
